compiler: move xtool to llgo/llgo, move xtool/env -> internal/env
