+++ go.mod
@@ -2,4 +2,11 @@ module github.com/goplus/llgo
 
 go 1.20
 
+require (
+	github.com/goplus/gogen v1.16.4
+	github.com/qiniu/x v1.13.10
+)
+
+require golang.org/x/tools v0.19.0 // indirect
+
 retract v0.8.0
