+++ xtool/ar/common.go
