+++ xtool/clang/types/scanner/scanner.go
